refactor(): moved optionParsers to api
